+++ tools/pdf_export/pdf_export/__init__.py
@@ -1,0 +1,5 @@
+"""PDF export package for the Plurality Wiki project."""
+
+from .cli import main
+
+__all__ = ["main"]
