+++ tools/pdf_export/pdf_export/__main__.py
@@ -1,0 +1,9 @@
+"""Allow running ``python -m pdf_export``."""
+
+from __future__ import annotations
+
+from .cli import main
+
+
+if __name__ == "__main__":
+    main()
