+++ tools/pdf_export/pdf_export/constants.py
@@ -1,0 +1,27 @@
+"""Module containing constant values used across the exporter."""
+
+from __future__ import annotations
+
+DEFAULT_PDF_ENGINES = [
+    "xelatex",
+    "tectonic",
+    "pdflatex",
+]
+
+CJK_FONT_CANDIDATES = [
+    "Noto Serif CJK SC",
+    "Noto Sans CJK SC",
+    "Source Han Serif SC",
+    "Source Han Sans SC",
+    "思源宋体",
+    "思源黑体",
+    "SimSun",
+    "SimHei",
+    "Microsoft YaHei",
+    "PingFang SC",
+    "Songti SC",
+]
+
+DEFAULT_COVER_TITLE = "plurality_wiki"
+DEFAULT_COVER_SUBTITLE = "多重意识体系统知识库"
+DEFAULT_COVER_FOOTER = "plurality_wiki 项目"
