+++ tools/pdf_export/pdf_export/requirements.py
@@ -1,0 +1,63 @@
+"""Runtime requirement helpers for the PDF exporter."""
+
+from __future__ import annotations
+
+import shutil
+import subprocess
+from typing import Iterable
+
+from .constants import CJK_FONT_CANDIDATES, DEFAULT_PDF_ENGINES
+
+
+def check_requirements(pandoc_cmd: str) -> None:
+    """Ensure that the external tools required by the script are available."""
+
+    if shutil.which(pandoc_cmd) is None:
+        raise SystemExit(
+            f"未找到 `{pandoc_cmd}` 可执行文件。请先安装 Pandoc 后再运行此脚本。"
+        )
+
+
+def _first_available(executables: Iterable[str]) -> str | None:
+    for candidate in executables:
+        if shutil.which(candidate):
+            return candidate
+    return None
+
+
+def detect_pdf_engine(preferred: str | None) -> str | None:
+    """Find an available PDF engine for Pandoc, if any."""
+
+    if preferred:
+        if shutil.which(preferred) is None:
+            raise SystemExit(
+                f"未找到指定的 PDF 引擎 `{preferred}`。请确认其已安装或改用其他引擎。"
+            )
+        return preferred
+
+    return _first_available(DEFAULT_PDF_ENGINES)
+
+
+def detect_cjk_font() -> str | None:
+    """Return the first available Chinese font family, if any."""
+
+    if shutil.which("fc-list") is None:
+        return None
+
+    for candidate in CJK_FONT_CANDIDATES:
+        try:
+            result = subprocess.run(
+                ["fc-list", candidate],
+                check=False,
+                text=True,
+                encoding="utf-8",
+                errors="replace",
+                capture_output=True,
+            )
+        except OSError:
+            return None
+
+        if result.stdout.strip():
+            return candidate
+
+    return None
